refactor(manager/npm): extract shared workspace matcher from bun
